backend/requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
passlib==1.7.4
mysql_connector_repackaged==0.3.1
